Renamed all command classes that were not compaint with our naming conventions.
